Further validation refinement. CRUD professional.

+++ database/payers.php
@@ -180,12 +180,7 @@ function editEntityPayerValuePerK($accountid, $identitypayer, $valueperk)
 function deleteEntityPayer($accountid, $identitypayer)
 {
     global $conn;
-    $stmt = $conn->prepare("DELETE FROM entitypayer WHERE identitypayer = :identitypayer
-                            AND (SELECT
-                                procedure.identitypayer
-                              FROM procedure, procedureaccount
-                              WHERE procedureaccount.idaccount = :accountid AND procedureaccount.idprocedure = procedure.idprocedure AND procedure.identitypayer = :identitypayer
-                            ) IS NULL");
+    $stmt = $conn->prepare("DELETE FROM entitypayer WHERE identitypayer = :identitypayer AND idaccount = :accountid");
     $stmt->execute(array("accountid" => $accountid, "identitypayer" => $identitypayer));
 
     return $stmt->rowCount() > 0;
@@ -194,12 +189,7 @@ function deleteEntityPayer($accountid, $identitypayer)
 function deletePrivatePayer($accountid, $idprivatepayer)
 {
     global $conn;
-    $stmt = $conn->prepare("DELETE FROM privatepayer WHERE idprivatepayer = :idprivatepayer
-                            AND (SELECT
-                                procedure.idprivatepayer
-                              FROM procedure, procedureaccount
-                              WHERE procedureaccount.idaccount = :accountid AND procedureaccount.idprocedure = procedure.idprocedure AND procedure.idprivatepayer = :idprivatepayer
-                            ) IS NULL");
+    $stmt = $conn->prepare("DELETE FROM privatepayer WHERE idprivatepayer = :idprivatepayer AND idaccount = :accountid");
     $stmt->execute(array("accountid" => $accountid, "idprivatepayer" => $idprivatepayer));
 
     return $stmt->rowCount() > 0;
